refactor fullstacktrace to seperate file in database layer
